+++ Changes.md
@@ -6,6 +6,7 @@ Features
 
 - ClosestPointSampler : Added a new node for sampling primitive variables from the closest point on a source primitive.
 - CurveSampler : Added a new node for sampling primitive variables from parametric positions on some source curves.
+- Viewer : Added overlay to Image views to control interactive renders where available (#3419).
 
 Improvements
 ------------